code/requirements.txt
streamlit
google-generativeai
groq
python-docx
python-dotenv